dispatcher: Add schedule_monitor to check task schedulability
